feat: implement main layout with bottom navigation and localization support
